Relocating the bluetooth classes.
