+++ Nemesis-Mod/src/bluetooth/CustomUuid.cpp
