+++ Nemesis-Mod/src/bluetooth/CustomUuid.h
